pyproject.toml
[project]
name = "project2-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "google-generativeai>=0.8.5",
    "peewee>=3.18.1",
]
